requirements.txt
streamlit
psycopg2-binary
requests
PyPDF2
python-dotenv